chore: move calculateBodyLength to it's own file
